Single-File: Add Null-Terminator in-memory json stream.
On Linux, the in-situ parser checks that the input stream is null-terminated once the root element is completely parsed.
Ideally, the parser should check the size of the input stream, but it currently doesn't take the size into consideration.
Instead of locally modifying RapidJson code, we simply add a null-terminator at the end of the stream.

In the current implementation, it is safe to add the null-terminator because bundle::info_t::config_t::map() maps the
enture single-file bundle, and there is content beyond all embeded files.

This update is not performed on Windows in order to avoid an unnecessary copy-on-write.